polkitbackend: Add new GKeyFile based backend
This change adds an alternative PolkitBackendAuthority with a much
simpler ahead-of-time backend authority based around GKeyFile formatted
text files.

This may be enabled instead of the existing JS backend as a compile time
option.

To ensure there are no upgrade issues while polkitd is still running,
the old default rule file is preserved, whilst the new rules have a
different suffix: ".keyrules".

The new keyfile format is not too dissimilar from the old pkla format,
but it does decouple from the multiple results that could yieled based
on state, such as "ResultActive". Instead, a rule may only have a single
match state, "Result". If the Action ID is matched, and a ResultInverse
is specified, then this value will be returned if the other conditions
aren't met.

This means that each rule in a file specifies all the required state,
such as:

 SubjectActive=true
 SubjectLocal=true

Note that it is possible to also match `SubjectActive=false`, as the
keyfile evaluation does not do any comparison on defaults. Rather, the
presence of a key in the rule body will cause the value of the key
to be tested against the PolicyContext. This allows the format to be
very flexible, providing very specific matching rules with inverse
conditions and results.

Wildcard matching is provided for `Actions` string array, to allow
a global rule to be installed based on condition matching. Additionally
one can use the `ActionContains` directive to emulate the old JS indexOf
functionality.

All in all, this new backend vastly simplifies the maintainence of PolKit,
by completely doing away with a JavaScript dependency that must be
repeatedly updated to the latest mozjs release, causing issues for
distribution integration. Additionally, the new backend is far lighter
than the JS backend, making polkit friendly once more for embedded
environments, nullifying the need for various incompatible forks and
"drop in replacements". However, to assist in supporting projects still
requiring JS rules, the default backend remains "js" for now.

NOTE: This is the first rebase of this behaviour, and some fields aren't
yet implemented in the format such as UNIX net groups, plus I need to write
up documentation on the new format, port the test suites, etc.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>